Feature or Bugfix: Bugfix Binary Source: No Signed-off-by: suwenxiang <suwenxiang@huawei.com>

+++ interfaces/kits/gfx_utils/style.h
@@ -177,7 +177,7 @@ public:
     uint8_t imageOpa_;
     /* text style */
     uint8_t textOpa_;
-    uint8_t font_;
+    uint16_t font_;
     int8_t lineSpace_;
     int16_t letterSpace_;
     int16_t lineHeight_;
